+++ src/main/java/com/vendingMachine/model/entity/Product.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Data
 @Table
@@ -18,14 +19,14 @@ public class Product {
     @Id()
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    
+
     private String name;
-    
-    private Double price;
-    
+
+    private BigDecimal price;
+
     private Integer quantity;
 
     public void restQuantity() {
-        quantity --;
+        quantity--;
     }
 }
