+++ src/main/java/com/vendingMachine/service/VendingMachineService.java
@@ -5,17 +5,18 @@ import com.vendingMachine.model.dto.PurchaseResponse;
 import com.vendingMachine.model.entity.Product;
 import com.vendingMachine.model.entity.Purchase;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface VendingMachineService {
-    
+
     List<Product> getProductList();
-    
+
     Purchase sumCoins(PurchaseRequest purchaseRequest);
 
     PurchaseResponse buyProcess(Integer idProduct, Integer idPurchase);
 
     Purchase getPurchaseById(Integer purchaseId);
 
-    Double getPurchaseRefund(Integer purchaseId);
+    BigDecimal getPurchaseRefund(Integer purchaseId);
 }
